Add calendar sync, Apple integration, production configs, and security fixes
Features:
- Add generic calendar sync schema migrations
- Add Apple CalDAV client integration
- Add Microsoft calendar client
- Add shadow/replay engine for sync
- Add raft consensus for audit replicator
- Add Google calendar client adapter
- Add production Helm charts for Kubernetes deployment
- Add production documentation (compliance, install, helm)
- Add Terraform modules for production environments

Security:
- Fix dashboard API client TypeScript errors (unused tenantId params)
- All tenant_id now passed via X-Tenant-ID header from interceptor

Other:
- Various backend service improvements
- Frontend calendar and conflict resolution components
- Production infrastructure configs

+++ backend/cmd/notifications-service/main.go
@@ -335,7 +335,7 @@ func generateRandomString(length int) string {
 	const chars = "abcdefghijklmnopqrstuvwxyz0123456789"
 	result := make([]byte, length)
 	for i := range result {
-		result[i] = chars[time.Now().UnixNano()%len(chars)]
+		result[i] = chars[time.Now().UnixNano()%int64(len(chars))]
 		time.Sleep(time.Nanosecond)
 	}
 	return string(result)
